# PERF: improve get_loc on unsorted, non-unique indexes