--- a/pandas/_libs/index_class_helper.pxi.in
+++ b/pandas/_libs/index_class_helper.pxi.in
@@ -74,7 +74,7 @@ cdef class {{name}}Engine(IndexEngine):
         if count > 1:
             return indexer
         if count == 1:
-            return found[0]
+            return int(found[0])
 
         raise KeyError(val)