@jreback 2018-02-06
seems to be failing 2.7 on windows and a gazillion warnings

```
 C:\projects\pandas\pandas\core\indexes\base.py:1726: DeprecationWarning: using a non-integer number instead of an integer will result in an error in the future
    return getitem(key)
  C:\projects\pandas\pandas\core\indexes\base.py:1726: DeprecationWarning: using a non-integer number instead of an integer will result in an error in the future
    return getitem(key)
  C:\projects\pandas\pandas\core\indexes\base.py:1726: DeprecationWarning: using a non-integer number instead of an integer will result in an error in the future
    return getitem(key)
  C:\projects\pandas\pandas\core\indexes\base.py:1726: DeprecationWarning: using a non-integer number instead of an integer will result in an error in the future
    return getitem(key)
  C:\projects\pandas\pandas\core\indexes\base.py:1726: DeprecationWarning: using a non-integer number instead of an integer will result in an error in the future
    return getitem(key)
```