@jreback 2018-02-06
also can you see if this closes other issues as well (a couple of refernces in OP)