@toobaz 2018-05-07
@jreback : fixed (wasn't that hard!), ping.

(I guess it shouldn't harm to include in 0.23.0)